+++ DIPimage_sources.cmake
@@ -35,6 +35,7 @@ dipimage/bopening.cpp
 dipimage/bpropagation.cpp
 dipimage/bskeleton.cpp
 dipimage/canny.cpp
+dipimage/ced.cpp
 dipimage/clip.cpp
 dipimage/closing.cpp
 dipimage/convolve.cpp