+++ pydip/filtering.cpp
@@ -102,8 +102,14 @@ void init_filtering( py::module& m ) {
           "in"_a, "percentile"_a, "kernel"_a = dip::Kernel{}, "boundaryCondition"_a = dip::StringArray{} );
    m.def( "NonMaximumSuppression", py::overload_cast< dip::Image const&, dip::Image const&, dip::Image const&, dip::String const& >( &dip::NonMaximumSuppression ),
           "gradmag"_a, "gradient"_a, "mask"_a = dip::Image{}, "mode"_a = dip::S::INTERPOLATE );
-   m.def( "PeronaMalik", py::overload_cast< dip::Image const&, dip::uint, dip::dfloat, dip::dfloat, dip::String >( &dip::PeronaMalik ),
+   m.def( "PeronaMalikDiffusion", py::overload_cast< dip::Image const&, dip::uint, dip::dfloat, dip::dfloat, dip::String const& >( &dip::PeronaMalikDiffusion ),
           "in"_a, "iterations"_a = 5, "K"_a = 10, "lambda"_a = 0.25, "g"_a = "Gauss" );
+   m.def( "GaussianAnisotropicDiffusion", py::overload_cast< dip::Image const&, dip::uint, dip::dfloat, dip::dfloat, dip::String const& >( &dip::GaussianAnisotropicDiffusion ),
+          "in"_a, "iterations"_a = 5, "K"_a = 10, "lambda"_a = 0.25, "g"_a = "Gauss" );
+   m.def( "RobustAnisotropicDiffusion", py::overload_cast< dip::Image const&, dip::uint, dip::dfloat, dip::dfloat >( &dip::RobustAnisotropicDiffusion ),
+          "in"_a, "iterations"_a = 5, "sigma"_a = 10, "lambda"_a = 0.25 );
+   m.def( "CoherenceEnhancingDiffusion", py::overload_cast< dip::Image const&, dip::dfloat, dip::dfloat, dip::uint, dip::StringSet const& >( &dip::CoherenceEnhancingDiffusion ),
+          "in"_a, "derivativeSigma"_a = 1, "regularizationSigma"_a = 3, "iterations"_a = 5, "flags"_a = dip::StringSet{} );
 
    // diplib/transform.h
    m.def( "FourierTransform", py::overload_cast< dip::Image const&, dip::StringSet const&, dip::BooleanArray const& >( &dip::FourierTransform ),